refactor/#14 Long -> Int로 수정

+++ app/src/main/java/com/sopt/uniqlo/data/productlist/service/ProductService.kt
@@ -12,6 +12,6 @@ interface ProductService {
 
     @GET("products/{productId}")
     suspend fun getProductDetail(
-        @Path("productId") productId: Long
+        @Path("productId") productId: Int
     ): BaseResponse<ProductDetailResponseDto>
 }
